+++ app/src/main/java/eu/kanade/presentation/util/ChapterNumberFormatter.kt
@@ -8,6 +8,6 @@ private val formatter = DecimalFormat(
     DecimalFormatSymbols().apply { decimalSeparator = '.' },
 )
 
-fun formatChapterNumber(chapterNumber: Float): String {
+fun formatChapterNumber(chapterNumber: Double): String {
     return formatter.format(chapterNumber)
 }
